fix subgraph tests
